+++ tests/ModelNestedArrayTest.php
@@ -86,7 +86,7 @@ class ModelNestedArrayTest extends TestCase
         ]);
         $mInner->removeField('_id');
         $mInner->idField = 'uid';
-        $mInner->addField('uid', ['actual' => '_id', 'type' => 'integer']);
+        $mInner->addField('uid', ['actual' => '_id', 'type' => 'bigint']);
         $mInner->addField('name');
         $mInner->addField('y', ['actual' => '_birthday', 'type' => 'date']);
         $mInner->addCondition('uid', '!=', 3);
